refactor(daemon): extract runner::output submodule (#2642)
Move truncate_output and brief-mode constants to runner/output.rs.
